+++ src/run.ts
@@ -13,11 +13,9 @@ export async function runApp(context: vscode.ExtensionContext) {
     // only to re-use a terminal if it's using the same interpreter.
     const pythonAPI =
       vscode.extensions.getExtension("ms-python.python")!.exports;
-    const pythonExecCommand = (
-      await pythonAPI.environment.getExecutionDetails(
-        vscode.window.activeTextEditor?.document.uri
-      )
-    ).execCommand.join(" ");
+    const pythonExecCommand = pythonAPI.environments.getActiveEnvironmentPath(
+      vscode.window.activeTextEditor?.document.uri
+    ).path;
 
     const shinyTerminals = vscode.window.terminals.filter(
       (term) => term.name === TERMINAL_NAME